requirements/requirements.txt
Django==5.0.2
gunicorn==21.2.0
pillow==10.2.0
django-cors-headers==4.3.1
django-ckeditor==6.5.0